feat(ostream): crude float/double support
Relates to, but does not substitute:
https://github.com/malachi-iot/estdlib/issues/23

+++ test/arduino/cmath/src/main.cpp
@@ -12,7 +12,7 @@ void setup()
 
 void loop()
 {
-    static unsigned counter;
+    static float counter;
     cout << F("sim(") << ++counter << F(")");
     cout << estd::endl;
